# swagger.json
{"openapi":"3.0.0","paths":{"/auth/register":{"post":{"operationId":"AuthController_register","summary":"","description":"","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterUserDto"}}}},"responses":{"201":{"description":""}},"tags":["auth"],"security":[{"bearer":[]}]}},"/auth/login":{"post":{"operationId":"AuthController_login","summary":"","description":"","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginUserDto"}}}},"responses":{"200":{"description":""}},"tags":["auth"],"security":[{"bearer":[]}]}},"/users/me":{"get":{"operationId":"UserController_getMe","summary":"","description":"","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["users"],"security":[{"bearer":[]}]}},"/users":{"get":{"operationId":"UserController_getUsers","summary":"","description":"","parameters":[{"name":"limit","required":false,"in":"query","schema":{"minimum":1,"maximum":50,"default":10,"type":"number"}},{"name":"page","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}}],"responses":{"200":{"description":""}},"tags":["users"],"security":[{"bearer":[]}]}},"/todo-lists":{"post":{"operationId":"TodoListController_createTodoList","summary":"","description":"","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTodoListDto"}}}},"responses":{"201":{"description":""}},"tags":["todo-lists"],"security":[{"bearer":[]}]}},"/todo-lists/{todoListId}":{"get":{"operationId":"TodoListController_getTodoList","summary":"","description":"","parameters":[{"name":"todoListId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["todo-lists"],"security":[{"bearer":[]}]},"patch":{"operationId":"TodoListController_updateTodoListName","summary":"","description":"","parameters":[{"name":"todoListId","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTodoListNameDto"}}}},"responses":{"200":{"description":""}},"tags":["todo-lists"],"security":[{"bearer":[]}]},"delete":{"operationId":"TodoListController_deleteTodoList","summary":"","description":"","parameters":[{"name":"todoListId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["todo-lists"],"security":[{"bearer":[]}]}},"/todo-lists/share":{"post":{"operationId":"TodoListController_shareTodoList","summary":"","description":"","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ShareTodoListDto"}}}},"responses":{"200":{"description":""}},"tags":["todo-lists"],"security":[{"bearer":[]}]}},"/todo-lists/{todoListId}/items":{"post":{"operationId":"ItemController_createTodoListItem","summary":"","description":"","parameters":[{"name":"todoListId","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTodoListItemDto"}}}},"responses":{"201":{"description":""}},"tags":["todo-lists/:todoListId/items"],"security":[{"bearer":[]}]}},"/todo-lists/{todoListId}/items/{todoListItemId}":{"patch":{"operationId":"ItemController_updateTodoListItemState","summary":"","description":"","parameters":[{"name":"todoListId","required":true,"in":"path","schema":{"type":"number"}},{"name":"todoListItemId","required":true,"in":"path","schema":{"type":"number"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateTodoListItemStateDto"}}}},"responses":{"200":{"description":""}},"tags":["todo-lists/:todoListId/items"],"security":[{"bearer":[]}]},"get":{"operationId":"ItemController_getTodoListItem","summary":"","description":"","parameters":[{"name":"todoListId","required":true,"in":"path","schema":{"type":"number"}},{"name":"todoListItemId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["todo-lists/:todoListId/items"],"security":[{"bearer":[]}]},"delete":{"operationId":"ItemController_deleteTodoListItem","summary":"","description":"","parameters":[{"name":"todoListId","required":true,"in":"path","schema":{"type":"number"}},{"name":"todoListItemId","required":true,"in":"path","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["todo-lists/:todoListId/items"],"security":[{"bearer":[]}]}}},"info":{"title":"ToDo app","description":"API for ToDo list management application","version":"1.0","contact":{}},"tags":[{"name":"todo","description":""}],"servers":[],"components":{"schemas":{"RegisterUserDto":{"type":"object","properties":{"firstName":{"type":"string"},"lastName":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"confirmPassword":{"type":"string"}},"required":["firstName","lastName","email","password","confirmPassword"]},"LoginUserDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"CreateTodoListDto":{"type":"object","properties":{"name":{"type":"string"}},"required":["name"]},"ShareTodoListDto":{"type":"object","properties":{"todoListId":{"type":"number"},"userId":{"type":"number"}},"required":["todoListId","userId"]},"UpdateTodoListNameDto":{"type":"object","properties":{"name":{"type":"string"}},"required":["name"]},"CreateTodoListItemDto":{"type":"object","properties":{"title":{"type":"string"},"description":{"type":"string"},"deadline":{"format":"date-time","type":"string"},"state":{"enum":["ACTIVE","COMPLETED","CANCELLED"],"type":"string"}},"required":["title","description","deadline","state"]},"UpdateTodoListItemStateDto":{"type":"object","properties":{"state":{"enum":["ACTIVE","COMPLETED","CANCELLED"],"type":"string"}},"required":["state"]}}}}